pyproject.toml
[project]
name = "civil-ai-project"
version = "0.1.0"
description = "建設法務 AI エージェント - ホスト実行スクリプト用環境"
requires-python = ">=3.11"
dependencies = [
    "SPARQLWrapper>=2.0.0",
    "neo4j>=6.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
]
